Update int64 doctest outputs for nanobind's numpy export
nanobind exports numpy arrays via DLPack, which encodes only bit-width
and signedness, not the C typedef identity. On Linux, scipp::index
(int64_t) is typedef'd to `long`, which pybind11's buffer-protocol
export mapped to numpy's platform-default int dtype (so repr() omitted
the dtype, e.g. `array([15, 15])`). nanobind instead always resolves
64-bit signed ints to the canonical `int64` dtype, which is a distinct
dtype object from the platform default even though both are 8 bytes,
so repr() now includes it explicitly (`array([15, 15], dtype=int64)`).

Values are unchanged; only the printed dtype label differs. Updated
the 16 affected doctest examples in arithmetic.py, bins.py, math.py
and shape.py to match, verified against the built nanobind module
with the same doctest flags (ELLIPSIS | NORMALIZE_WHITESPACE) Sphinx
uses.

+++ src/scipp/core/arithmetic.py
@@ -125,7 +125,7 @@ def floor_divide(dividend: VariableLike, divisor: VariableLike) -> VariableLike:
     --------
 
     >>> sc.floor_divide(sc.arange('x', -2, 3), sc.scalar(2)).values
-    array([-1, -1,  0,  0,  1])
+    array([-1, -1,  0,  0,  1], dtype=int64)
     >>> sc.floor_divide(sc.arange('x', -2.0, 3.0), sc.scalar(2.0)).values
     array([-1., -1.,  0.,  0.,  1.])
 
@@ -183,14 +183,14 @@ def mod(dividend: VariableLike, divisor: VariableLike) -> VariableLike:
     --------
 
     >>> sc.mod(sc.arange('x', -3, 5), sc.scalar(3)).values
-    array([0, 1, 2, 0, 1, 2, 0, 1])
+    array([0, 1, 2, 0, 1, 2, 0, 1], dtype=int64)
     >>> sc.mod(sc.arange('x', -3, 5), sc.scalar(-3)).values
-    array([ 0, -2, -1,  0, -2, -1,  0, -2])
+    array([ 0, -2, -1,  0, -2, -1,  0, -2], dtype=int64)
 
     Or equivalently in operator notation
 
     >>> (sc.arange('x', -3, 5) % sc.scalar(3)).values
-    array([0, 1, 2, 0, 1, 2, 0, 1])
+    array([0, 1, 2, 0, 1, 2, 0, 1], dtype=int64)
 
     See Also
     --------
